box appearing on one screen
