+++ src/App.css
@@ -36,3 +36,19 @@
 .Bold {
   font-weight: bold;
 }
+
+.playing-controls img {
+  margin: 10px;
+  cursor: pointer;
+}
+
+.playing-controls {
+  padding: 30px 0;
+  display: flex;
+  justify-content: center;
+}
+
+.up-down-arrows {
+  display: flex;
+  flex-direction: column;
+}
